Abstract prompt maxi parts into smaller components
