+++ src/components/prompt-control/components/modify-tab/index.js
@@ -3,19 +3,14 @@
  */
 import { __ } from '@wordpress/i18n';
 import { useEffect, useState } from '@wordpress/element';
-import { MediaUpload, MediaUploadCheck } from '@wordpress/block-editor';
 
 /**
  * Internal dependencies
  */
-import {
-	Button,
-	DialogBox,
-	ReactSelectControl,
-	TextareaControl,
-} from '../../../../components';
-import ResultCard from '../results-card';
-import { downloadTextFile } from '../../../../editor/style-cards/utils';
+import { Button } from '../../../../components';
+import HistoryButtons from '../history-buttons';
+import ResultCards from '../result-cards';
+import ResultsTopBar from '../results-top-bar';
 import {
 	getContentAttributesSection,
 	getContextSection,
@@ -25,11 +20,6 @@ import {
 	handleContentGeneration,
 } from '../../utils';
 import { MODIFICATION_ACTIONS, MODIFY_OPTIONS } from '../../constants';
-
-/**
- * External dependencies
- */
-import { capitalize, isEmpty } from 'lodash';
 
 /**
  * Styles
@@ -60,7 +50,6 @@ const ModifyTab = ({
 }) => {
 	const [modifyOption, setModifyOption] = useState(MODIFY_OPTIONS[0]);
 	const [customText, setCustomText] = useState('');
-	const [loadUntilIndex, setLoadUntilIndex] = useState(5);
 
 	useEffect(() => {
 		if (!selectedResultId) {
@@ -155,27 +144,10 @@ Your task is to maintain the original intent and context while ${modificationAct
 		});
 	};
 
-	const handleHistorySelect = media => {
-		fetch(media.url)
-			// Need to parse the response 2 times,
-			// because it was stringified twice in the export function
-			.then(response => response.json())
-			.then(response => JSON.parse(response))
-			.then(jsonData => {
-				setResults(jsonData);
-			})
-			.catch(error => {
-				console.error(error);
-			});
-	};
-
-	const handleHistoryExport = () => {
-		downloadTextFile(results, 'history.txt');
-	};
-
 	const cleanHistory = () => {
 		setResults([]);
 		setHistoryStartId(null);
+		setSelectedResultId(null);
 		switchToGenerateTab();
 	};
 
@@ -183,224 +155,42 @@ Your task is to maintain the original intent and context while ${modificationAct
 
 	return (
 		<div className={className}>
-			<div className={`${className}__buttons`}>
-				<MediaUploadCheck>
-					<MediaUpload
-						onSelect={handleHistorySelect}
-						allowedTypes='text'
-						render={({ open }) => (
-							<Button
-								className={`${className}__button`}
-								onClick={open}
-							>
-								{__('Import history', 'maxi-blocks')}
-							</Button>
-						)}
-					/>
-				</MediaUploadCheck>
-				<Button
-					className={`${className}__button`}
-					onClick={handleHistoryExport}
-				>
-					{__('Export history', 'maxi-blocks')}
-				</Button>
-			</div>
-			<div className={`${className}__top-bar`}>
-				{(!isEmpty(results) || selectedResultId) && (
-					<div className={`${className}__modification-options`}>
-						<ReactSelectControl
-							value={{
-								label: __(
-									capitalize(modifyOption),
-									'maxi-blocks'
-								),
-								value: modifyOption,
-							}}
-							defaultValue={{
-								label: __(
-									capitalize(modifyOption),
-									'maxi-blocks'
-								),
-								value: modifyOption,
-							}}
-							options={MODIFY_OPTIONS.map(option => ({
-								label: __(capitalize(option), 'maxi-blocks'),
-								value: option,
-							}))}
-							onChange={({ value }) => setModifyOption(value)}
-							isDisabled={isEmpty(results)}
-						/>
-						<Button
-							className={`${className}__button`}
-							onClick={modifyContent}
-							disabled={isEmpty(results) && !selectedResultId}
-						>
-							{__('Go!', 'maxi-blocks')}
-						</Button>
-					</div>
-				)}
-				{modifyOption === 'custom' && (
-					<TextareaControl
-						className={`${className}__textarea`}
-						placeholder={__(
-							'More information gives better results',
-							'maxi-blocks'
-						)}
-						value={customText}
-						onChange={val => setCustomText(val)}
-						disableResize
-					/>
-				)}
-				<div className={`${className}__buttons`}>
-					{results.every(result => !result.isSelectedText) && (
-						<Button
-							className={`${className}__button`}
-							onClick={switchToGenerateTab}
-						>
-							{__('Back', 'maxi-blocks')}
-						</Button>
-					)}
-					{!isEmpty(results) && (
-						<DialogBox
-							message={__(
-								'Are you sure you want to clean the history?',
-								'maxi-blocks'
-							)}
-							cancelLabel={__('Cancel', 'maxi-blocks')}
-							confirmLabel={__('Clean', 'maxi-blocks')}
-							onConfirm={cleanHistory}
-							buttonClassName={`${className}__button ${className}__clean-history-button`}
-							buttonChildren={__('Clean history', 'maxi-blocks')}
-						/>
-					)}
-				</div>
-			</div>
-			<div className={`${className}__results`}>
-				{selectedText && (
-					<ResultCard
-						result={{
-							content: selectedText,
-							isSelectedText: true,
-						}}
-						isSelected={selectedResultId === 'selectedText'}
-						isCustom={modifyOption === 'custom'}
-						onSelect={() => setSelectedResultId('selectedText')}
-					/>
-				)}
-				{results.map((result, index) => {
-					if (index >= loadUntilIndex) {
-						return null;
-					}
-
-					const refResultIndex =
-						result.refId &&
-						results.findIndex(
-							refResult => refResult.id === result.refId
-						);
-
-					const handleResultInsertion = () => {
-						const { start: selectionStart, end: selectionEnd } =
-							formatValue;
-
-						const contentBeforeSelection = content.slice(
-							0,
-							selectionStart
-						);
-						const contentAfterSelection =
-							content.slice(selectionEnd);
-						const newContent =
-							contentBeforeSelection +
-							result.content +
-							contentAfterSelection;
-
-						if (selectionStart !== selectionEnd) {
-							onChangeTextFormat({
-								...formatValue,
-								end: selectionStart + result.content.length,
-							});
-						}
-
-						onContentChange(newContent);
-					};
-
-					const handleResultSelection = (id = result.id) =>
-						setSelectedResultId(id);
-
-					const handleResultUseSettings = () => {
-						if (result.refId) {
-							setModifyOption(result.modificationType);
-							setCustomText(result.customText);
-
-							if (refResultIndex >= loadUntilIndex) {
-								setLoadUntilIndex(refResultIndex + 1);
-							}
-
-							setSelectedResultId(result.refId);
-
-							return;
-						}
-
-						if (result.settings) {
-							updateSettings(result.settings);
-							switchToGenerateTab();
-						}
-					};
-
-					const handleResultDeletion = () => {
-						if (historyStartId === result.id) {
-							setHistoryStartId(historyStartId - 1);
-						}
-
-						setResults(prevResults => {
-							const newResults = [...prevResults].filter(
-								deletedResult => deletedResult.id !== result.id
-							);
-
-							if (isEmpty(newResults)) {
-								switchToGenerateTab();
-							}
-
-							return newResults;
-						});
-
-						if (selectedResultId === result.id) {
-							setSelectedResultId(null);
-						}
-					};
-
-					return (
-						<ResultCard
-							// eslint-disable-next-line react/no-array-index-key
-							key={index}
-							index={index + 1}
-							result={result}
-							isFromPreviousSession={result.id <= historyStartId}
-							isSelected={result.id === selectedResultId}
-							isSelectedText={!!selectedText}
-							isCustom={modifyOption === 'custom'}
-							isRefExist={refResultIndex && refResultIndex >= 0}
-							onInsert={handleResultInsertion}
-							onSelect={handleResultSelection}
-							onUseSettings={handleResultUseSettings}
-							onDelete={handleResultDeletion}
-						/>
-					);
-				})}
-			</div>
-			{isEmpty(results) && !selectedResultId && (
-				<h4 className={`${className}__no-results`}>
-					{__('History is empty', 'maxi-blocks')}
-				</h4>
-			)}
-			{results.length > loadUntilIndex && (
-				<div className={`${className}__load-more`}>
-					<Button
-						onClick={() => setLoadUntilIndex(loadUntilIndex + 5)}
-					>
-						{__('Load more', 'maxi-blocks')}
-					</Button>
-				</div>
-			)}
+			<HistoryButtons
+				className={`${className}__buttons`}
+				buttonClassName={`${className}__button`}
+				results={results}
+				setResults={setResults}
+				setSelectedResultId={setSelectedResultId}
+			/>
+			<ResultsTopBar
+				results={results}
+				selectedResultId={selectedResultId}
+				modifyOption={modifyOption}
+				modifyContent={modifyContent}
+				customText={customText}
+				cleanHistory={cleanHistory}
+				setModifyOption={setModifyOption}
+				setCustomText={setCustomText}
+				switchToGenerateTab={switchToGenerateTab}
+			/>
+			<ResultCards
+				results={results}
+				selectedText={selectedText}
+				content={content}
+				modifyOption={modifyOption}
+				formatValue={formatValue}
+				historyStartId={historyStartId}
+				selectedResultId={selectedResultId}
+				setResults={setResults}
+				setSelectedResultId={setSelectedResultId}
+				setModifyOption={setModifyOption}
+				setCustomText={setCustomText}
+				setHistoryStartId={setHistoryStartId}
+				onContentChange={onContentChange}
+				onChangeTextFormat={onChangeTextFormat}
+				updateSettings={updateSettings}
+				switchToGenerateTab={switchToGenerateTab}
+			/>
 			{isGenerating && (
 				<Button className={`${className}__abort`} onClick={onAbort}>
 					{__('Stop', 'maxi-blocks')}
