+++ src/components/prompt-control/components/result-card/editor.scss
