+++ src/components/prompt-control/components/result-card/index.js
