+++ src/components/prompt-control/components/results-top-bar/editor.scss
@@ -1,0 +1,16 @@
+.maxi-prompt-control-results-top-bar {
+	top: 47px;
+	position: sticky;
+	padding: 8px 0;
+	background: var(--maxi-white-4-color);
+
+	&__modification-options {
+		display: flex;
+		gap: 8px;
+
+		& > *:first-child {
+			flex-grow: 1;
+		}
+	}
+
+}
